requirements.txt
streamlit==1.35.0
langchain==0.2.1
langchain-community==0.2.1
langchain-google-genai==1.0.7
chromadb==0.5.0
pypdf==4.2.0
python-dotenv==1.0.1
google-generativeai==0.7.0